refactor: combine home page and refactor contexts
