+++ electron/package.json
@@ -25,19 +25,20 @@
     "devDependencies": {
         "@electron-forge/cli": "^7.4.0",
         "@electron-forge/maker-deb": "^7.4.0",
+        "@electron-forge/maker-dmg": "^7.4.0",
         "@electron-forge/maker-rpm": "^7.4.0",
         "@electron-forge/maker-squirrel": "^7.4.0",
         "@electron-forge/maker-zip": "^7.4.0",
         "@electron-forge/plugin-auto-unpack-natives": "^7.4.0",
         "@electron-forge/plugin-fuses": "^7.4.0",
         "@electron-forge/plugin-vite": "^7.4.0",
-        "@electron-forge/maker-dmg": "^7.4.0",
         "@iconify/react": "^5.0.1",
         "@types/react": "^18.3.3",
         "@types/react-dom": "^18.3.0",
         "@typescript-eslint/eslint-plugin": "^5.0.0",
         "@typescript-eslint/parser": "^5.0.0",
         "autoprefixer": "^10.4.19",
+        "electron": "31.0.0",
         "eslint": "^8.0.1",
         "eslint-import-resolver-alias": "^1.1.2",
         "eslint-import-resolver-typescript": "^3.6.1",
@@ -47,14 +48,12 @@
         "ts-node": "^10.0.0",
         "typescript": "~4.5.4",
         "vite": "^5.0.12",
-        "vite-plugin-static-copy": "^1.0.5",
-        "electron": "31.0.0"
+        "vite-plugin-static-copy": "^1.0.5"
     },
     "keywords": [],
     "license": "AGPL-3.0-only",
     "dependencies": {
         "@electron/fuses": "^1.8.0",
-        "electron": "31.0.0",
         "@monaco-editor/react": "^4.6.0",
         "@radix-ui/react-accordion": "^1.1.2",
         "@radix-ui/react-checkbox": "^1.0.4",
@@ -75,6 +74,7 @@
         "axios": "^1.7.2",
         "class-variance-authority": "^0.7.0",
         "cmdk": "^1.0.0",
+        "electron": "31.0.0",
         "electron-debug": "^4.0.0",
         "electron-is-dev": "^3.0.1",
         "electron-log": "^5.1.5",
