+++ electron/src/frontend/components/panels/editor/components/file-tabs/action-item.tsx
