+++ electron/src/frontend/components/panels/editor/components/file-tabs/file-tabs.tsx
